set up auth as action for decoy controller
