+++ src/components/adminDecoy/adminDecoy.module.css
@@ -1,0 +1,153 @@
+.container {
+    margin-top: 5%;
+    height: 70%;
+    width: 85%;
+    font-family: sans-serif;
+    margin-left: auto;
+    margin-right: auto;
+    background-image: linear-gradient(to bottom, rgba(70, 188, 6, 0.856), rgb(6, 110, 60) 100%);
+    color: black;
+    border-radius: 10px;
+    border: solid 1px, black;
+    text-align: center;
+    justify-content: center;
+    /* display:flex; */
+}
+
+.dashboardContainer {
+    margin-top: 5%;
+    height: 70%;
+    width: 85%;
+    font-family: sans-serif;
+    margin-left: auto;
+    margin-right: auto;
+    background-image: linear-gradient(to bottom, rgba(70, 188, 6, 0.856), rgb(6, 110, 60) 100%);
+    color: black;
+    border-radius: 10px;
+    border: solid 1px, black;
+    text-align: center;
+    justify-content: center;
+    display:flex;
+
+}
+
+.title {
+    font-size: 30px;
+    padding-top: 1%;
+}
+
+.image {    
+    margin-top: 1em; 
+    width: 30%; 
+    height:20%;   
+}
+
+.dataContainer {
+    background-color: rgba(7, 116, 112, 0.764);
+    margin-right: 1%;
+    margin-left: 1%;
+    display: flex;
+    height: 85%;
+    border-radius: 10px;
+    overflow: auto;
+    scrollbar-width: thin;
+    box-shadow: 1px 1px 1px rgba(79, 78, 78, 0.633);
+}
+
+.contentContainer {
+    background-color: rgba(7, 116, 112, 0.764);
+    margin:auto;
+    width: 95%;
+    height: 90%;
+    border-radius: 10px;
+    overflow: auto;
+    scrollbar-width: thin;
+    box-shadow: 1px 1px 1px rgba(79, 78, 78, 0.633);
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.dataColumn { 
+    height: fit-content;
+    padding: 1%;
+    gap:20px;
+    display:flex;
+    flex-direction: row;
+    flex-wrap: wrap;   
+    justify-content: center;
+    background-color: rgba(36, 12, 145, 0.764);    
+    /* max-width: 800px; */
+
+}
+.data {
+    background-image: linear-gradient(to bottom, rgba(70, 188, 6, 0.856), rgb(6, 110, 60) 100%);
+    justify-content: center;
+    border-radius: 10px;
+    padding: 2%;
+    box-shadow: 1px 1px 1px rgba(78, 79, 78, 0.633);
+    /* margin-top: 2%; */
+    width:200px;
+}
+
+
+
+.alertOptions{
+    width: 80%;
+    display:flex;
+    flex-direction: column;
+    margin-left: auto;
+    margin-right: auto;
+}
+
+.par {
+    margin-top: 2%;
+    margin-bottom: 2%;
+
+}
+
+.button{
+    min-width: 200px;
+}
+/* @media only screen and (max-width: 800px) {
+    .container {
+        height: 90%;
+    }
+    .dashboardContainer {
+        height: 90%;
+    }
+
+    .dataContainer {
+        flex-direction: column;
+        justify-content: center;
+        align-content: center;
+    }
+    .image {
+        width: auto;
+        height: 10%;
+    }
+
+    .dataContainer> :nth-child(2)> :nth-child(1) {
+        margin-top: 0%;
+        background-color: red;
+    }
+
+    .dataColumn {
+        width: 90%;
+    }
+
+
+}
+
+ */
+
+ /* @media only screen and (max-width: 1000px){
+    .image {    
+        margin-top: 1em; 
+        width: 60%; 
+        height:auto;   
+    }
+ }
+ */
+
+
